When calculating size, the last column and last row should not have spacing added
The width of the layout should not include spacing after the last
column and after the last row. Besides a wrong width, all columns will
be the spacing amount shifted to the left. The increased height will
result in empty space at the bottom.

This commit only includes fix for horizontal layout but a similar
problem must exist for vertical layout.